--- sources/core_api/src/main/java/ru/mertsalovda/core_api/mediator/AppWithFacade.kt
@@ -1,8 +1,0 @@
-package ru.mertsalovda.core_api.mediator
-
-import ru.mertsalovda.core_api.providers.ProvidersFacade
-
-interface AppWithFacade {
-
-    fun getFacade(): ProvidersFacade
-}
