+++ sources/core_api/src/main/java/ru/mertsalovda/core_api/providers/AppWithFacade.kt
@@ -1,0 +1,6 @@
+package ru.mertsalovda.core_api.providers
+
+interface AppWithFacade {
+
+    fun getFacade(): ProvidersFacade
+}
