+++ sources/core_impl/src/main/java/ru/mertsalovda/core_impl/database/DatabaseComponent.kt
@@ -1,4 +1,4 @@
-package ru.mertsalovda.core_impl
+package ru.mertsalovda.core_impl.database
 
 import dagger.Component
 import ru.mertsalovda.core_api.database.DatabaseProvider
